requirements.txt
praw==7.8.1
transformers==4.48.3
torch==2.6.0
python-dotenv==1.0.1
